feat: migrate tests to layer 1 and add fixtures
