+++ contracts/package.json
@@ -25,8 +25,6 @@
     "test:parallel": "npx hardhat test test/unitTests/**/*.ts test/unitTests/*/**/*.ts test/unitTests/*/*/**/*.ts --parallel",
     "test:coverage": "npm run test:coverage:layer1 && npm run test:coverage:layer2 && npm run test:coverage:layer3 && npm run test:coverage:factory && npm run test:coverage:resolver && npm run test:coverage:resolverProxy && npm run test:timeTravel",
     "test:coverage:layer1": "npx hardhat coverage --testfiles 'test/unitTests/layer_1/**/*.ts'",
-    "test:coverage:layer2": "npx hardhat coverage --testfiles 'test/unitTests/layer_2/**/*.ts'",
-    "test:coverage:layer3": "npx hardhat coverage --testfiles 'test/unitTests/layer_3/**/*.ts'",
     "test:coverage:factory": "npx hardhat coverage --testfiles 'test/unitTests/factory/**/*.ts'",
     "test:coverage:resolver": "npx hardhat coverage --testfiles 'test/unitTests/resolver/**/*.ts'",
     "test:coverage:resolverProxy": "npx hardhat coverage --testfiles 'test/unitTests/resolverProxy/**/*.ts'",
@@ -38,23 +36,23 @@
     "test:accessControl": "npx hardhat test test/unitTests/layer_1/accessControl/accessControl.test.ts",
     "test:controlList": "npx hardhat test test/unitTests/layer_1/controlList/controlList.test.ts",
     "test:pause": "npx hardhat test test/unitTests/layer_1/pause/pause.test.ts",
-    "test:corporateActions": "npx hardhat test test/unitTests/layer_2/corporateActions/corporateActions.test.ts",
+    "test:corporateActions": "npx hardhat test test/unitTests/layer_1/corporateActions/corporateActions.test.ts",
     "test:erc1643": "npx hardhat test test/unitTests/layer_1/ERC1400/ERC1643/erc1643.test.ts",
     "test:erc1644": "npx hardhat test test/unitTests/layer_1/ERC1400/ERC1644/erc1644.test.ts",
     "test:erc1594": "npx hardhat test test/unitTests/layer_1/ERC1400/ERC1594/erc1594.test.ts",
-    "test:erc1400": "npx hardhat test test/unitTests/layer_2/ERC1400/erc1400.test.ts",
+    "test:erc1410": "npx hardhat test test/unitTests/layer_1/ERC1400/ERC1410/erc1410.test.ts",
     "test:erc20": "npx hardhat test test/unitTests/layer_1/ERC1400/ERC20/erc20.test.ts",
     "test:snapshots": "npx hardhat test test/unitTests/layer_1/snapshots/snapshots.test.ts",
-    "test:Snapshots": "npx hardhat test test/unitTests/layer_2/snapshots/snapshots.test.ts",
-    "test:adjustBalances": "npx hardhat test test/unitTests/layer_2/adjustBalances/adjustBalances.test.ts",
-    "test:equity": "npx hardhat test test/unitTests/layer_2/equity/equity.test.ts",
-    "test:bond": "npx hardhat test test/unitTests/layer_2/bond/bond.test.ts",
-    "test:scheduledSnapshots": "npx hardhat test test/unitTests/layer_2/scheduledTasks/scheduledSnapshots/scheduledSnapshots.test.ts",
-    "test:scheduledBalanceAdjustments": "npx hardhat test test/unitTests/layer_2/scheduledTasks/scheduledBalanceAdjustments/scheduledBalanceAdjustments.test.ts",
-    "test:scheduledTasks": "npx hardhat test test/unitTests/layer_2/scheduledTasks/scheduledTasks/scheduledTasks.test.ts",
+    "test:adjustBalances": "npx hardhat test test/unitTests/layer_1/adjustBalances/adjustBalances.test.ts",
+    "test:equity": "npx hardhat test test/unitTests/layer_1/equity/equity.test.ts",
+    "test:bond": "npx hardhat test test/unitTests/layer_1/bond/bond.test.ts",
+    "test:scheduledSnapshots": "npx hardhat test test/unitTests/layer_1/scheduledTasks/scheduledSnapshots/scheduledSnapshots.test.ts",
+    "test:scheduledBalanceAdjustments": "npx hardhat test test/unitTests/layer_1/scheduledTasks/scheduledBalanceAdjustments/scheduledBalanceAdjustments.test.ts",
+    "test:scheduledTasks": "npx hardhat test test/unitTests/layer_1/scheduledTasks/scheduledTasks/scheduledTasks.test.ts",
     "test:cap": "npx hardhat test test/unitTests/layer_1/cap/cap.test.ts",
     "test:Cap": "npx hardhat test test/unitTests/layer_2/cap/cap.test.ts",
-    "test:securityUSA": "npx hardhat test test/unitTests/layer_3/securityUSA/securityUSA.test.ts",
+    "test:securityUSA": "npx hardhat test test/unitTests/layer_1/securityUSA/securityUSA.test.ts",
+    "test:transferAndLock": "npx hardhat test test/unitTests/layer_1/transferAndLock/transferAndLock.test.ts",
     "test:lock": "npx hardhat test test/unitTests/layer_1/lock/lock.test.ts",
     "test:Lock": "npx hardhat test test/unitTests/layer_2/lock/lock.test.ts",
     "test:hold": "npx hardhat test test/unitTests/layer_1/hold/hold.test.ts",
