+++ contracts/test/unitTests/layer_1/bond/bond.test.ts
@@ -205,80 +205,77 @@
 
 import { expect } from 'chai'
 import { ethers } from 'hardhat'
+import { BigNumber } from 'ethers'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers.js'
 import { isinGenerator } from '@thomaschaplin/isin-generator'
 import {
-    type ResolverProxy,
-    type EquityUSA,
-    type Pause,
-    type AccessControl,
-    TimeTravel,
+    ResolverProxy,
+    BondUSA,
+    AccessControl,
+    Pause,
     Lock,
     Hold,
+    TimeTravel,
     ERC1410ScheduledTasks,
     IFactory,
     BusinessLogicResolver,
-    AccessControl__factory,
-    EquityUSATimeTravel__factory,
-    Pause__factory,
+    ERC1410ScheduledTasks__factory,
     Lock__factory,
     Hold__factory,
-    ERC1410ScheduledTasks__factory,
+    Pause__factory,
+    AccessControl__factory,
+    BondUSATimeTravel__factory,
     TimeTravel__factory,
     Kyc,
     SsiManagement,
 } from '@typechain'
 import {
     CORPORATE_ACTION_ROLE,
-    DEFAULT_PARTITION,
-    ISSUER_ROLE,
-    LOCKER_ROLE,
     PAUSER_ROLE,
+    BOND_MANAGER_ROLE,
+    LOCKER_ROLE,
+    ISSUER_ROLE,
     KYC_ROLE,
     SSI_MANAGER_ROLE,
-    deployEquityFromFactory,
+    DEFAULT_PARTITION,
     Rbac,
+    deployBondFromFactory,
     RegulationSubType,
     RegulationType,
     deployAtsFullInfrastructure,
     DeployAtsFullInfrastructureCommand,
     ADDRESS_ZERO,
+    MAX_UINT256,
+    ZERO,
+    EMPTY_STRING,
 } from '@scripts'
 import { grantRoleAndPauseToken } from '../../../common'
 import { dateToUnixTimestamp } from '../../../dateFormatter'
+import { loadFixture } from '@nomicfoundation/hardhat-network-helpers'
 
-const DECIMALS = 7
-let dividendsRecordDateInSeconds = 0
-let dividendsExecutionDateInSeconds = 0
-const dividendsAmountPerEquity = 1
-
-let votingRecordDateInSeconds = 0
+const numberOfUnits = 1000
+let startingDate = 0
+const numberOfCoupons = 50
+const frequency = 7
+const rate = 1
+let maturityDate = 0
+let firstCouponDate = 0
 const countriesControlListType = true
 const listOfCountries = 'ES,FR,CH'
 const info = 'info'
 
-let balanceAdjustmentExecutionDateInSeconds = 0
-const balanceAdjustmentFactor = 356
-const balanceAdjustmentDecimals = 2
+let couponRecordDateInSeconds = 0
+let couponExecutionDateInSeconds = 0
+const couponRate = 5
+const EMPTY_VC_ID = EMPTY_STRING
 
-const voteData = '0x'
-let votingData = {
-    recordDate: votingRecordDateInSeconds.toString(),
-    data: voteData,
+let couponData = {
+    recordDate: couponRecordDateInSeconds.toString(),
+    executionDate: couponExecutionDateInSeconds.toString(),
+    rate: couponRate,
 }
-let dividendData = {
-    recordDate: dividendsRecordDateInSeconds.toString(),
-    executionDate: dividendsExecutionDateInSeconds.toString(),
-    amount: dividendsAmountPerEquity,
-}
-let balanceAdjustmentData = {
-    executionDate: balanceAdjustmentExecutionDateInSeconds.toString(),
-    factor: balanceAdjustmentFactor,
-    decimals: balanceAdjustmentDecimals,
-}
-const number_Of_Shares = 100000n
 
-describe('Equity Tests', () => {
+describe('Bond Tests', () => {
     let diamond: ResolverProxy
     let signer_A: SignerWithAddress
     let signer_B: SignerWithAddress
@@ -290,7 +287,7 @@ describe('Equity Tests', () => {
 
     let factory: IFactory
     let businessLogicResolver: BusinessLogicResolver
-    let equityFacet: EquityUSA
+    let bondFacet: BondUSA
     let accessControlFacet: AccessControl
     let pauseFacet: Pause
     let lockFacet: Lock
@@ -299,6 +296,91 @@ describe('Equity Tests', () => {
     let timeTravelFacet: TimeTravel
     let kycFacet: Kyc
     let ssiManagementFacet: SsiManagement
+
+    function set_initRbacs(): Rbac[] {
+        const rbacPause: Rbac = {
+            role: PAUSER_ROLE,
+            members: [account_B],
+        }
+        const rbacKYC: Rbac = {
+            role: KYC_ROLE,
+            members: [account_B],
+        }
+        const rbacSSI: Rbac = {
+            role: SSI_MANAGER_ROLE,
+            members: [account_A],
+        }
+        return [rbacPause, rbacKYC, rbacSSI]
+    }
+
+    async function setFacets({ diamond }: { diamond: ResolverProxy }) {
+        bondFacet = BondUSATimeTravel__factory.connect(
+            diamond.address,
+            signer_A
+        )
+        accessControlFacet = AccessControl__factory.connect(
+            diamond.address,
+            signer_A
+        )
+        pauseFacet = Pause__factory.connect(diamond.address, signer_A)
+        lockFacet = Lock__factory.connect(diamond.address, signer_A)
+        holdFacet = Hold__factory.connect(diamond.address, signer_A)
+        erc1410Facet = ERC1410ScheduledTasks__factory.connect(
+            diamond.address,
+            signer_A
+        )
+        timeTravelFacet = TimeTravel__factory.connect(diamond.address, signer_A)
+        kycFacet = await ethers.getContractAt('Kyc', diamond.address, signer_B)
+        ssiManagementFacet = await ethers.getContractAt(
+            'SsiManagement',
+            diamond.address,
+            signer_A
+        )
+
+        await ssiManagementFacet.connect(signer_A).addIssuer(account_A)
+        await kycFacet.grantKyc(
+            account_A,
+            EMPTY_VC_ID,
+            ZERO,
+            MAX_UINT256,
+            account_A
+        )
+    }
+
+    async function deploySecurityFixtureSinglePartition() {
+        let init_rbacs: Rbac[] = set_initRbacs()
+
+        diamond = await deployBondFromFactory({
+            adminAccount: account_A,
+            isWhiteList: false,
+            isControllable: true,
+            arePartitionsProtected: false,
+            clearingActive: false,
+            isMultiPartition: false,
+            name: 'TEST_AccessControl',
+            symbol: 'TAC',
+            decimals: 6,
+            isin: isinGenerator(),
+            currency: '0x455552',
+            numberOfUnits,
+            nominalValue: 100,
+            startingDate,
+            maturityDate,
+            couponFrequency: frequency,
+            couponRate: rate,
+            firstCouponDate,
+            regulationType: RegulationType.REG_D,
+            regulationSubType: RegulationSubType.REG_D_506_C,
+            countriesControlListType,
+            listOfCountries,
+            info,
+            init_rbacs,
+            factory,
+            businessLogicResolver: businessLogicResolver.address,
+        })
+
+        await setFacets({ diamond })
+    }
 
     before(async () => {
         // mute | mock console.log
@@ -324,122 +406,36 @@ describe('Equity Tests', () => {
     })
 
     beforeEach(async () => {
-        const rbacPause: Rbac = {
-            role: PAUSER_ROLE,
-            members: [account_B],
+        startingDate = dateToUnixTimestamp(`2030-01-01T00:00:35Z`)
+        maturityDate = startingDate + numberOfCoupons * frequency
+        firstCouponDate = startingDate + 1
+        couponRecordDateInSeconds = dateToUnixTimestamp(`2030-01-01T00:01:00Z`)
+        couponExecutionDateInSeconds =
+            dateToUnixTimestamp(`2030-01-01T00:10:00Z`)
+        couponData = {
+            recordDate: couponRecordDateInSeconds.toString(),
+            executionDate: couponExecutionDateInSeconds.toString(),
+            rate: couponRate,
         }
-        const rbacKYC: Rbac = {
-            role: KYC_ROLE,
-            members: [account_B],
-        }
-        const rbacSSI: Rbac = {
-            role: SSI_MANAGER_ROLE,
-            members: [account_A],
-        }
-        const init_rbacs: Rbac[] = [rbacPause, rbacKYC, rbacSSI]
-
-        diamond = await deployEquityFromFactory({
-            adminAccount: account_A,
-            isWhiteList: false,
-            isControllable: true,
-            arePartitionsProtected: false,
-            clearingActive: false,
-            isMultiPartition: false,
-            name: 'TEST_AccessControl',
-            symbol: 'TAC',
-            decimals: DECIMALS,
-            isin: isinGenerator(),
-            votingRight: false,
-            informationRight: false,
-            liquidationRight: false,
-            subscriptionRight: true,
-            conversionRight: true,
-            redemptionRight: true,
-            putRight: false,
-            dividendRight: 1,
-            currency: '0x345678',
-            numberOfShares: number_Of_Shares,
-            nominalValue: 100,
-            regulationType: RegulationType.REG_D,
-            regulationSubType: RegulationSubType.REG_D_506_B,
-            countriesControlListType,
-            listOfCountries,
-            info,
-            init_rbacs,
-            businessLogicResolver: businessLogicResolver.address,
-            factory,
-        })
-
-        accessControlFacet = AccessControl__factory.connect(
-            diamond.address,
-            signer_A
-        )
-        equityFacet = EquityUSATimeTravel__factory.connect(
-            diamond.address,
-            signer_A
-        )
-        pauseFacet = Pause__factory.connect(diamond.address, signer_A)
-        lockFacet = Lock__factory.connect(diamond.address, signer_A)
-        holdFacet = Hold__factory.connect(diamond.address, signer_A)
-        erc1410Facet = ERC1410ScheduledTasks__factory.connect(
-            diamond.address,
-            signer_A
-        )
-        timeTravelFacet = TimeTravel__factory.connect(diamond.address, signer_A)
-        rbacSSI
-        kycFacet = await ethers.getContractAt('Kyc', diamond.address, signer_B)
-        ssiManagementFacet = await ethers.getContractAt(
-            'SsiManagement',
-            diamond.address,
-            signer_A
-        )
-
-        await ssiManagementFacet.connect(signer_A).addIssuer(account_A)
-        await kycFacet.grantKyc(account_A, '', 0, 9999999999, account_A)
-
-        dividendsRecordDateInSeconds = dateToUnixTimestamp(
-            '2030-01-01T00:00:10Z'
-        )
-        dividendsExecutionDateInSeconds = dateToUnixTimestamp(
-            '2030-01-01T00:16:40Z'
-        )
-        votingRecordDateInSeconds = dateToUnixTimestamp('2030-01-01T00:00:10Z')
-        balanceAdjustmentExecutionDateInSeconds = dateToUnixTimestamp(
-            '2030-01-01T00:00:10Z'
-        )
-
-        votingData = {
-            recordDate: votingRecordDateInSeconds.toString(),
-            data: voteData,
-        }
-        dividendData = {
-            recordDate: dividendsRecordDateInSeconds.toString(),
-            executionDate: dividendsExecutionDateInSeconds.toString(),
-            amount: dividendsAmountPerEquity,
-        }
-        balanceAdjustmentData = {
-            executionDate: balanceAdjustmentExecutionDateInSeconds.toString(),
-            factor: balanceAdjustmentFactor,
-            decimals: balanceAdjustmentDecimals,
-        }
+        await loadFixture(deploySecurityFixtureSinglePartition)
     })
 
     afterEach(async () => {
-        await timeTravelFacet.resetSystemTimestamp()
+        timeTravelFacet.resetSystemTimestamp()
     })
 
-    describe('Dividends', () => {
-        it('GIVEN an account without corporateActions role WHEN setDividends THEN transaction fails with AccountHasNoRole', async () => {
+    describe('Coupons', () => {
+        it('GIVEN an account without corporateActions role WHEN setCoupon THEN transaction fails with AccountHasNoRole', async () => {
             // Using account C (non role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
 
-            // set dividend fails
-            await expect(
-                equityFacet.setDividends(dividendData)
-            ).to.be.rejectedWith('AccountHasNoRole')
+            // set coupon fails
+            await expect(bondFacet.setCoupon(couponData)).to.be.rejectedWith(
+                'AccountHasNoRole'
+            )
         })
 
-        it('GIVEN a paused Token WHEN setDividends THEN transaction fails with TokenIsPaused', async () => {
+        it('GIVEN a paused Token WHEN setCoupon THEN transaction fails with TokenIsPaused', async () => {
             // Granting Role to account C and Pause
             await grantRoleAndPauseToken(
                 accessControlFacet,
@@ -451,111 +447,106 @@ describe('Equity Tests', () => {
             )
 
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
 
-            // set dividend fails
-            await expect(
-                equityFacet.setDividends(dividendData)
-            ).to.be.rejectedWith('TokenIsPaused')
-        })
-
-        it('GIVEN an account with corporateActions role WHEN setDividends with wrong dates THEN transaction fails', async () => {
-            await timeTravelFacet.changeSystemTimestamp(
-                dateToUnixTimestamp('2030-01-01T00:00:00Z')
+            // set coupon fails
+            await expect(bondFacet.setCoupon(couponData)).to.be.rejectedWith(
+                'TokenIsPaused'
             )
-            // Granting Role to account C
-            accessControlFacet = accessControlFacet.connect(signer_A)
-            await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
-            // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
-
-            // set dividend
-            const wrongDividendData_1 = {
-                recordDate: dividendsExecutionDateInSeconds.toString(),
-                executionDate: dividendsRecordDateInSeconds.toString(),
-                amount: dividendsAmountPerEquity,
-            }
-
-            await expect(
-                equityFacet.setDividends(wrongDividendData_1)
-            ).to.be.revertedWithCustomError(equityFacet, 'WrongDates')
-
-            const wrongDividendData_2 = {
-                recordDate: dateToUnixTimestamp(
-                    '2029-12-31T23:59:59Z'
-                ).toString(),
-                executionDate: dividendsExecutionDateInSeconds.toString(),
-                amount: dividendsAmountPerEquity,
-            }
-
-            await expect(
-                equityFacet.setDividends(wrongDividendData_2)
-            ).to.be.revertedWithCustomError(equityFacet, 'WrongTimestamp')
         })
 
-        it('GIVEN an account with corporateActions role WHEN setDividends THEN transaction succeeds', async () => {
+        it('GIVEN an account with corporateActions role WHEN setCoupon with wrong dates THEN transaction fails', async () => {
             // Granting Role to account C
             accessControlFacet = accessControlFacet.connect(signer_A)
             await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
 
-            // set dividend
-            await expect(equityFacet.setDividends(dividendData))
-                .to.emit(equityFacet, 'DividendSet')
+            // set coupon
+            const wrongcouponData_1 = {
+                recordDate: couponExecutionDateInSeconds.toString(),
+                executionDate: couponRecordDateInSeconds.toString(),
+                rate: couponRate,
+            }
+
+            await expect(
+                bondFacet.setCoupon(wrongcouponData_1)
+            ).to.be.revertedWithCustomError(bondFacet, 'WrongDates')
+
+            const wrongcouponData_2 = {
+                recordDate: (
+                    (await ethers.provider.getBlock('latest')).timestamp - 1
+                ).toString(),
+                executionDate: couponExecutionDateInSeconds.toString(),
+                rate: couponRate,
+            }
+
+            await expect(
+                bondFacet.setCoupon(wrongcouponData_2)
+            ).to.be.revertedWithCustomError(bondFacet, 'WrongTimestamp')
+        })
+
+        it('GIVEN an account with corporateActions role WHEN setCoupon THEN transaction succeeds', async () => {
+            // Granting Role to account C
+            accessControlFacet = accessControlFacet.connect(signer_A)
+            await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
+            // Using account C (with role)
+            bondFacet = bondFacet.connect(signer_C)
+
+            // set coupon
+            await expect(bondFacet.setCoupon(couponData))
+                .to.emit(bondFacet, 'CouponSet')
                 .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
+                    '0x0000000000000000000000000000000000000000000000000000000000000033',
+                    numberOfCoupons + 1,
                     account_C,
-                    dividendsRecordDateInSeconds,
-                    dividendsExecutionDateInSeconds,
-                    dividendsAmountPerEquity
+                    couponRecordDateInSeconds,
+                    couponExecutionDateInSeconds,
+                    couponRate
                 )
 
             // check list members
-            await expect(equityFacet.getDividends(1000)).to.be.rejectedWith(
+            await expect(bondFacet.getCoupon(1000)).to.be.rejectedWith(
                 'WrongIndexForAction'
             )
 
-            const listCount = await equityFacet.getDividendsCount()
-            const dividend = await equityFacet.getDividends(1)
-            const dividendFor = await equityFacet.getDividendsFor(1, account_A)
+            const listCount = await bondFacet.getCouponCount()
+            const coupon = await bondFacet.getCoupon(numberOfCoupons + 1)
+            const couponFor = await bondFacet.getCouponFor(
+                numberOfCoupons + 1,
+                account_A
+            )
 
-            expect(listCount).to.equal(1)
-            expect(dividend.snapshotId).to.equal(0)
-            expect(dividend.dividend.recordDate).to.equal(
-                dividendsRecordDateInSeconds
+            expect(listCount).to.equal(numberOfCoupons + 1)
+            expect(coupon.snapshotId).to.equal(0)
+            expect(coupon.coupon.recordDate).to.equal(couponRecordDateInSeconds)
+            expect(coupon.coupon.executionDate).to.equal(
+                couponExecutionDateInSeconds
             )
-            expect(dividend.dividend.executionDate).to.equal(
-                dividendsExecutionDateInSeconds
+            expect(coupon.coupon.rate).to.equal(couponRate)
+            expect(couponFor.recordDate).to.equal(couponRecordDateInSeconds)
+            expect(couponFor.executionDate).to.equal(
+                couponExecutionDateInSeconds
             )
-            expect(dividend.dividend.amount).to.equal(dividendsAmountPerEquity)
-            expect(dividendFor.recordDate).to.equal(
-                dividendsRecordDateInSeconds
-            )
-            expect(dividendFor.executionDate).to.equal(
-                dividendsExecutionDateInSeconds
-            )
-            expect(dividendFor.amount).to.equal(dividendsAmountPerEquity)
-            expect(dividendFor.tokenBalance).to.equal(0)
-            expect(dividendFor.recordDateReached).to.equal(false)
-            expect(dividendFor.decimals).to.equal(0)
+            expect(couponFor.rate).to.equal(couponRate)
+            expect(couponFor.tokenBalance).to.equal(0)
+            expect(couponFor.recordDateReached).to.equal(false)
         })
 
-        it('GIVEN an account with corporateActions role WHEN setDividends and lock THEN transaction succeeds', async () => {
+        it('GIVEN an account with corporateActions role WHEN setCoupon and lock THEN transaction succeeds', async () => {
             // Granting Role to account C
             accessControlFacet = accessControlFacet.connect(signer_A)
             await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
             await accessControlFacet.grantRole(LOCKER_ROLE, account_C)
             await accessControlFacet.grantRole(ISSUER_ROLE, account_C)
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
             lockFacet = lockFacet.connect(signer_C)
             erc1410Facet = erc1410Facet.connect(signer_C)
 
             // issue and lock
-            const TotalAmount = number_Of_Shares
-            const LockedAmount = TotalAmount - 5n
+            const TotalAmount = numberOfUnits
+            const LockedAmount = TotalAmount - 5
 
             await erc1410Facet.issueByPartition({
                 partition: DEFAULT_PARTITION,
@@ -564,42 +555,47 @@ describe('Equity Tests', () => {
                 data: '0x',
             })
 
-            await lockFacet.lock(LockedAmount, account_A, 99999999999)
+            await lockFacet.lock(LockedAmount, account_A, MAX_UINT256)
 
-            // set dividend
-            await expect(equityFacet.setDividends(dividendData))
-                .to.emit(equityFacet, 'DividendSet')
+            // set coupon
+            await expect(bondFacet.setCoupon(couponData))
+                .to.emit(bondFacet, 'CouponSet')
                 .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
+                    '0x0000000000000000000000000000000000000000000000000000000000000033',
+                    numberOfCoupons + 1,
                     account_C,
-                    dividendsRecordDateInSeconds,
-                    dividendsExecutionDateInSeconds,
-                    dividendsAmountPerEquity
+                    couponRecordDateInSeconds,
+                    couponExecutionDateInSeconds,
+                    couponRate
                 )
 
             // check list members
             await timeTravelFacet.changeSystemTimestamp(
-                dividendsRecordDateInSeconds + 1
+                couponRecordDateInSeconds + 1
             )
-            const dividendFor = await equityFacet.getDividendsFor(1, account_A)
+            await accessControlFacet.revokeRole(ISSUER_ROLE, account_C)
 
-            expect(dividendFor.tokenBalance).to.equal(TotalAmount)
-            expect(dividendFor.recordDateReached).to.equal(true)
+            const couponFor = await bondFacet.getCouponFor(
+                numberOfCoupons + 1,
+                account_A
+            )
+
+            expect(couponFor.tokenBalance).to.equal(TotalAmount)
+            expect(couponFor.recordDateReached).to.equal(true)
         })
 
-        it('GIVEN an account with corporateActions role WHEN setDividends and hold THEN transaction succeeds', async () => {
+        it('GIVEN an account with corporateActions role WHEN setCoupon and hold THEN transaction succeeds', async () => {
             // Granting Role to account C
             accessControlFacet = accessControlFacet.connect(signer_A)
             await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
             await accessControlFacet.grantRole(ISSUER_ROLE, account_C)
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
             erc1410Facet = erc1410Facet.connect(signer_C)
 
             // issue and hold
-            const TotalAmount = number_Of_Shares
-            const HeldAmount = TotalAmount - 5n
+            const TotalAmount = numberOfUnits
+            const HeldAmount = TotalAmount - 5
 
             await erc1410Facet.issueByPartition({
                 partition: DEFAULT_PARTITION,
@@ -610,7 +606,7 @@ describe('Equity Tests', () => {
 
             let hold = {
                 amount: HeldAmount,
-                expirationTimestamp: 999999999999999,
+                expirationTimestamp: MAX_UINT256,
                 escrow: account_B,
                 to: ADDRESS_ZERO,
                 data: '0x',
@@ -618,206 +614,182 @@ describe('Equity Tests', () => {
 
             await holdFacet.createHoldByPartition(DEFAULT_PARTITION, hold)
 
-            // set dividend
-            await expect(equityFacet.setDividends(dividendData))
-                .to.emit(equityFacet, 'DividendSet')
+            // set coupon
+            await expect(bondFacet.setCoupon(couponData))
+                .to.emit(bondFacet, 'CouponSet')
                 .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
+                    '0x0000000000000000000000000000000000000000000000000000000000000033',
+                    numberOfCoupons + 1,
                     account_C,
-                    dividendsRecordDateInSeconds,
-                    dividendsExecutionDateInSeconds,
-                    dividendsAmountPerEquity
+                    couponRecordDateInSeconds,
+                    couponExecutionDateInSeconds,
+                    couponRate
                 )
 
             // check list members
             await timeTravelFacet.changeSystemTimestamp(
-                dividendsRecordDateInSeconds + 1
+                couponRecordDateInSeconds + 1
             )
-            const dividendFor = await equityFacet.getDividendsFor(1, account_A)
+            await accessControlFacet.revokeRole(ISSUER_ROLE, account_C)
 
-            expect(dividendFor.tokenBalance).to.equal(TotalAmount)
-            expect(dividendFor.recordDateReached).to.equal(true)
-        })
-    })
-
-    describe('Voting rights', () => {
-        it('GIVEN an account without corporateActions role WHEN setVoting THEN transaction fails with AccountHasNoRole', async () => {
-            // Using account C (non role)
-            equityFacet = equityFacet.connect(signer_C)
-
-            // set dividend fails
-            await expect(equityFacet.setVoting(votingData)).to.be.rejectedWith(
-                'AccountHasNoRole'
+            const couponFor = await bondFacet.getCouponFor(
+                numberOfCoupons + 1,
+                account_A
             )
+
+            expect(couponFor.tokenBalance).to.equal(TotalAmount)
+            expect(couponFor.recordDateReached).to.equal(true)
         })
 
-        it('GIVEN a paused Token WHEN setVoting THEN transaction fails with TokenIsPaused', async () => {
-            // Granting Role to account C and Pause
-            await grantRoleAndPauseToken(
-                accessControlFacet,
-                pauseFacet,
-                CORPORATE_ACTION_ROLE,
-                signer_A,
-                signer_B,
-                account_C
-            )
-
-            // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
-
-            // set dividend fails
-            await expect(equityFacet.setVoting(votingData)).to.be.rejectedWith(
-                'TokenIsPaused'
-            )
-        })
-
-        it('GIVEN an account with corporateActions role WHEN setVoting THEN transaction succeeds', async () => {
+        it('GIVEN an account with bondManager role WHEN setMaturityDate THEN transaction succeeds', async () => {
+            // * Arrange
             // Granting Role to account C
             accessControlFacet = accessControlFacet.connect(signer_A)
-            await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
+            await accessControlFacet.grantRole(BOND_MANAGER_ROLE, account_C)
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
+            // Get maturity date
+            const maturityDateBefore = (await bondFacet.getBondDetails())
+                .maturityDate
+            // New maturity date
+            const newMaturityDate = maturityDateBefore.add(
+                BigNumber.from(86400)
+            )
 
-            // set dividend
-            await expect(equityFacet.setVoting(votingData))
-                .to.emit(equityFacet, 'VotingSet')
+            // * Act
+            // Set maturity date
+            const receipt = await bondFacet.updateMaturityDate(newMaturityDate)
+
+            // * Assert
+            await expect(receipt)
+                .to.emit(bondFacet, 'MaturityDateUpdated')
                 .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
-                    account_C,
-                    votingRecordDateInSeconds,
-                    voteData
+                    bondFacet.address,
+                    newMaturityDate,
+                    maturityDateBefore
                 )
-
-            // check list members
-            // await expect(equityFacet.getVoting(1000)).to.be.rejectedWith(
-            //     'WrongIndexForAction'
-            // )
-
-            const listCount = await equityFacet.getVotingCount()
-            const voting = await equityFacet.getVoting(1)
-            const votingFor = await equityFacet.getVotingFor(1, account_A)
-
-            expect(listCount).to.equal(1)
-            expect(voting.snapshotId).to.equal(0)
-            expect(voting.voting.recordDate).to.equal(votingRecordDateInSeconds)
-            expect(voting.voting.data).to.equal(voteData)
-            expect(votingFor.recordDate).to.equal(dividendsRecordDateInSeconds)
-            expect(votingFor.data).to.equal(voteData)
-            expect(votingFor.tokenBalance).to.equal(0)
-            expect(votingFor.recordDateReached).to.equal(false)
+            // check date
+            const maturityDateAfter = (await bondFacet.getBondDetails())
+                .maturityDate
+            expect(maturityDateAfter).not.to.be.equal(maturityDateBefore)
+            expect(maturityDateAfter).to.be.equal(newMaturityDate)
         })
 
-        it('GIVEN an account with corporateActions role WHEN setVoting and lock THEN transaction succeeds', async () => {
+        it('GIVEN an account with bondManager role WHEN setMaturityDate to earlier date THEN transaction fails', async () => {
+            // * Arrange
             // Granting Role to account C
             accessControlFacet = accessControlFacet.connect(signer_A)
-            await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
-            await accessControlFacet.grantRole(LOCKER_ROLE, account_C)
-            await accessControlFacet.grantRole(ISSUER_ROLE, account_C)
+            await accessControlFacet.grantRole(BOND_MANAGER_ROLE, account_C)
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
-            lockFacet = lockFacet.connect(signer_C)
-            erc1410Facet = erc1410Facet.connect(signer_C)
-
-            // issue and lock
-            const TotalAmount = number_Of_Shares
-            const LockedAmount = TotalAmount - 5n
-
-            await erc1410Facet.issueByPartition({
-                partition: DEFAULT_PARTITION,
-                tokenHolder: account_A,
-                value: TotalAmount,
-                data: '0x',
-            })
-            await lockFacet.lock(LockedAmount, account_A, 99999999999)
-
-            // set dividend
-            await expect(equityFacet.setVoting(votingData))
-                .to.emit(equityFacet, 'VotingSet')
-                .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
-                    account_C,
-                    votingRecordDateInSeconds,
-                    voteData
-                )
-
-            await timeTravelFacet.changeSystemTimestamp(
-                votingRecordDateInSeconds + 1
+            bondFacet = bondFacet.connect(signer_C)
+            // Get maturity date
+            const maturityDateBefore = (await bondFacet.getBondDetails())
+                .maturityDate
+            // New maturity date (earlier than current)
+            // New maturity date (earlier than current)
+            const dayBeforeCurrentMaturity = maturityDateBefore.sub(
+                BigNumber.from(86400)
             )
-            const votingFor = await equityFacet.getVotingFor(1, account_A)
 
-            expect(votingFor.tokenBalance).to.equal(TotalAmount)
-            expect(votingFor.recordDateReached).to.equal(true)
-        })
-    })
-
-    describe('Balance adjustments', () => {
-        it('GIVEN an account without corporateActions role WHEN setBalanceAdjustment THEN transaction fails with AccountHasNoRole', async () => {
-            // Using account C (non role)
-            equityFacet = equityFacet.connect(signer_C)
-
-            // set dividend fails
+            // * Act & Assert
+            // Set maturity date
             await expect(
-                equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData)
+                bondFacet.updateMaturityDate(dayBeforeCurrentMaturity)
+            ).to.be.revertedWithCustomError(bondFacet, 'BondMaturityDateWrong')
+            // Ensure maturity date is not updated
+            const maturityDateAfter = (await bondFacet.getBondDetails())
+                .maturityDate
+            expect(maturityDateAfter).to.be.equal(maturityDateBefore)
+        })
+
+        it('GIVEN an account without bondManager role WHEN setMaturityDate THEN transaction fails with AccountHasNoRole', async () => {
+            // * Arrange
+            // Using account C (without role)
+            bondFacet = bondFacet.connect(signer_C)
+            // Get maturity date
+            const maturityDateBefore = (await bondFacet.getBondDetails())
+                .maturityDate
+            // New maturity date
+            const newMaturityDate = maturityDateBefore.add(
+                BigNumber.from(86400)
+            )
+
+            // * Act & Assert
+            // Set maturity date
+            await expect(
+                bondFacet.updateMaturityDate(newMaturityDate)
             ).to.be.rejectedWith('AccountHasNoRole')
+            // Ensure maturity date is not updated
+            const maturityDateAfter = (await bondFacet.getBondDetails())
+                .maturityDate
+            expect(maturityDateAfter).to.be.equal(maturityDateBefore)
         })
 
-        it('GIVEN a paused Token WHEN setBalanceAdjustment THEN transaction fails with TokenIsPaused', async () => {
+        it('GIVEN a paused Token WHEN setMaturityDate THEN transaction fails with TokenIsPaused', async () => {
+            // * Arrange
             // Granting Role to account C and Pause
             await grantRoleAndPauseToken(
                 accessControlFacet,
                 pauseFacet,
-                CORPORATE_ACTION_ROLE,
+                BOND_MANAGER_ROLE,
                 signer_A,
                 signer_B,
                 account_C
             )
-
             // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+            bondFacet = bondFacet.connect(signer_C)
+            // Get maturity date
+            const maturityDateBefore = (await bondFacet.getBondDetails())
+                .maturityDate
+            // New maturity date
+            const newMaturityDate = maturityDateBefore.add(
+                BigNumber.from(86400)
+            )
 
-            // set dividend fails
+            // * Act & Assert
+            // Set maturity date
             await expect(
-                equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData)
+                bondFacet.updateMaturityDate(newMaturityDate)
             ).to.be.rejectedWith('TokenIsPaused')
+            // Ensure maturity date is not updated
+            const maturityDateAfter = (await bondFacet.getBondDetails())
+                .maturityDate
+            expect(maturityDateAfter).to.be.equal(maturityDateBefore)
         })
 
-        it('GIVEN an account with corporateActions role WHEN setBalanceAdjustment THEN transaction succeeds', async () => {
-            // Granting Role to account C
-            accessControlFacet = accessControlFacet.connect(signer_A)
-            await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
-            // Using account C (with role)
-            equityFacet = equityFacet.connect(signer_C)
+        it('Check number of created Coupon', async () => {
+            bondFacet = bondFacet.connect(signer_C)
 
-            // set dividend
-            await expect(
-                equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData)
-            )
-                .to.emit(equityFacet, 'ScheduledBalanceAdjustmentSet')
-                .withArgs(
-                    '0x0000000000000000000000000000000000000000000000000000000000000001',
-                    1,
-                    account_C,
-                    balanceAdjustmentExecutionDateInSeconds,
-                    balanceAdjustmentFactor,
-                    balanceAdjustmentDecimals
+            const couponCount = await bondFacet.getCouponCount()
+
+            expect(couponCount).to.equal(numberOfCoupons)
+        })
+
+        it('Check Coupon', async () => {
+            bondFacet = bondFacet.connect(signer_C)
+
+            for (let i = 1; i <= numberOfCoupons; i++) {
+                const coupon = await bondFacet.getCoupon(i)
+                const couponFor = await bondFacet.getCouponFor(i, account_A)
+
+                expect(coupon.coupon.recordDate).to.equal(
+                    firstCouponDate + (i - 1) * frequency
                 )
-
-            const listCount =
-                await equityFacet.getScheduledBalanceAdjustmentCount()
-            const balanceAdjustment =
-                await equityFacet.getScheduledBalanceAdjustment(1)
-
-            expect(listCount).to.equal(1)
-            expect(balanceAdjustment.executionDate).to.equal(
-                balanceAdjustmentExecutionDateInSeconds
-            )
-            expect(balanceAdjustment.factor).to.equal(balanceAdjustmentFactor)
-            expect(balanceAdjustment.decimals).to.equal(
-                balanceAdjustmentDecimals
-            )
+                expect(coupon.coupon.executionDate).to.equal(
+                    firstCouponDate + (i - 1) * frequency
+                )
+                expect(coupon.coupon.rate).to.equal(rate)
+                expect(coupon.snapshotId).to.equal(0)
+                expect(couponFor.recordDate).to.equal(
+                    firstCouponDate + (i - 1) * frequency
+                )
+                expect(couponFor.executionDate).to.equal(
+                    firstCouponDate + (i - 1) * frequency
+                )
+                expect(couponFor.tokenBalance).to.equal(0)
+                expect(couponFor.rate).to.equal(rate)
+                expect(couponFor.recordDateReached).to.equal(false)
+            }
         })
     })
 })
