+++ contracts/test/unitTests/layer_1/scheduledTasks/scheduledSnapshots/scheduledSnapshots.test.ts
@@ -210,16 +210,16 @@ import { isinGenerator } from '@thomaschaplin/isin-generator'
 import {
     type ResolverProxy,
     type Equity,
-    type ScheduledBalanceAdjustments,
+    type ScheduledSnapshots,
     type AccessControl,
     ScheduledTasks,
     TimeTravel,
-    IFactory,
     BusinessLogicResolver,
+    IFactory,
     AccessControl__factory,
     Equity__factory,
-    ScheduledBalanceAdjustments__factory,
     ScheduledTasks__factory,
+    ScheduledSnapshots__factory,
     TimeTravel__factory,
 } from '@typechain'
 import {
@@ -234,8 +234,9 @@ import {
     MAX_UINT256,
 } from '@scripts'
 import { dateToUnixTimestamp } from '../../../../dateFormatter'
+import { loadFixture } from '@nomicfoundation/hardhat-network-helpers'
 
-describe('Scheduled BalanceAdjustments Tests', () => {
+describe('Scheduled Snapshots Tests', () => {
     let diamond: ResolverProxy
     let signer_A: SignerWithAddress
     let signer_B: SignerWithAddress
@@ -248,40 +249,13 @@ describe('Scheduled BalanceAdjustments Tests', () => {
     let factory: IFactory
     let businessLogicResolver: BusinessLogicResolver
     let equityFacet: Equity
-    let scheduledBalanceAdjustmentsFacet: ScheduledBalanceAdjustments
+    let scheduledSnapshotsFacet: ScheduledSnapshots
     let scheduledTasksFacet: ScheduledTasks
     let accessControlFacet: AccessControl
     let timeTravelFacet: TimeTravel
 
-    before(async () => {
-        // mute | mock console.log
-        console.log = () => {}
-        // eslint-disable-next-line @typescript-eslint/no-extra-semi
-        ;[signer_A, signer_B, signer_C] = await ethers.getSigners()
-        account_A = signer_A.address
-        account_B = signer_B.address
-        account_C = signer_C.address
-
-        const { deployer, ...deployedContracts } =
-            await deployAtsFullInfrastructure(
-                await DeployAtsFullInfrastructureCommand.newInstance({
-                    signer: signer_A,
-                    useDeployed: false,
-                    useEnvironment: false,
-                    timeTravelEnabled: true,
-                })
-            )
-
-        factory = deployedContracts.factory.contract
-        businessLogicResolver = deployedContracts.businessLogicResolver.contract
-    })
-
-    beforeEach(async () => {
-        const rbacPause: Rbac = {
-            role: PAUSER_ROLE,
-            members: [account_B],
-        }
-        const init_rbacs: Rbac[] = [rbacPause]
+    async function deploySecurityFixtureSinglePartition() {
+        const init_rbacs: Rbac[] = set_initRbacs()
 
         diamond = await deployEquityFromFactory({
             adminAccount: account_A,
@@ -290,7 +264,7 @@ describe('Scheduled BalanceAdjustments Tests', () => {
             arePartitionsProtected: false,
             clearingActive: false,
             isMultiPartition: false,
-            name: 'TEST_AccessControl',
+            name: 'TestScheduledSnapshots',
             symbol: 'TAC',
             decimals: 6,
             isin: isinGenerator(),
@@ -315,182 +289,195 @@ describe('Scheduled BalanceAdjustments Tests', () => {
             factory,
         })
 
+        await setFacets(diamond)
+    }
+
+    async function setFacets(diamond: ResolverProxy) {
         accessControlFacet = AccessControl__factory.connect(
             diamond.address,
             signer_A
         )
         equityFacet = Equity__factory.connect(diamond.address, signer_A)
-        scheduledBalanceAdjustmentsFacet =
-            ScheduledBalanceAdjustments__factory.connect(
-                diamond.address,
-                signer_A
-            )
+        scheduledSnapshotsFacet = ScheduledSnapshots__factory.connect(
+            diamond.address,
+            signer_A
+        )
         scheduledTasksFacet = ScheduledTasks__factory.connect(
             diamond.address,
             signer_A
         )
         timeTravelFacet = TimeTravel__factory.connect(diamond.address, signer_A)
+    }
+
+    function set_initRbacs(): Rbac[] {
+        const rbacPause: Rbac = {
+            role: PAUSER_ROLE,
+            members: [account_B],
+        }
+        return [rbacPause]
+    }
+
+    before(async () => {
+        // mute | mock console.log
+        console.log = () => {}
+        // eslint-disable-next-line @typescript-eslint/no-extra-semi
+        ;[signer_A, signer_B, signer_C] = await ethers.getSigners()
+        account_A = signer_A.address
+        account_B = signer_B.address
+        account_C = signer_C.address
+
+        const { deployer, ...deployedContracts } =
+            await deployAtsFullInfrastructure(
+                await DeployAtsFullInfrastructureCommand.newInstance({
+                    signer: signer_A,
+                    useDeployed: false,
+                    useEnvironment: true,
+                    timeTravelEnabled: true,
+                })
+            )
+
+        factory = deployedContracts.factory.contract
+        businessLogicResolver = deployedContracts.businessLogicResolver.contract
+    })
+
+    beforeEach(async () => {
+        await loadFixture(deploySecurityFixtureSinglePartition)
     })
 
     afterEach(async () => {
         timeTravelFacet.resetSystemTimestamp()
     })
 
-    it('GIVEN a token WHEN triggerBalanceAdjustments THEN transaction succeeds', async () => {
+    it('GIVEN a token WHEN triggerSnapshots THEN transaction succeeds', async () => {
         // Granting Role to account C
         accessControlFacet = accessControlFacet.connect(signer_A)
         await accessControlFacet.grantRole(CORPORATE_ACTION_ROLE, account_C)
         // Using account C (with role)
         equityFacet = equityFacet.connect(signer_C)
 
-        // set balanceAdjustment
-        const balanceAdjustmentExecutionDateInSeconds_1 = dateToUnixTimestamp(
+        // set dividend
+        const dividendsRecordDateInSeconds_1 = dateToUnixTimestamp(
             '2030-01-01T00:00:06Z'
         )
-        const balanceAdjustmentExecutionDateInSeconds_2 = dateToUnixTimestamp(
+        const dividendsRecordDateInSeconds_2 = dateToUnixTimestamp(
             '2030-01-01T00:00:12Z'
         )
-        const balanceAdjustmentExecutionDateInSeconds_3 = dateToUnixTimestamp(
+        const dividendsRecordDateInSeconds_3 = dateToUnixTimestamp(
             '2030-01-01T00:00:18Z'
         )
-        const balanceAdjustmentsFactor = 1
-        const balanceAdjustmentsDecimals = 2
+        const dividendsExecutionDateInSeconds = dateToUnixTimestamp(
+            '2030-01-01T00:01:00Z'
+        )
+        const dividendsAmountPerEquity = 1
+        const dividendData_1 = {
+            recordDate: dividendsRecordDateInSeconds_1.toString(),
+            executionDate: dividendsExecutionDateInSeconds.toString(),
+            amount: dividendsAmountPerEquity,
+        }
+        const dividendData_2 = {
+            recordDate: dividendsRecordDateInSeconds_2.toString(),
+            executionDate: dividendsExecutionDateInSeconds.toString(),
+            amount: dividendsAmountPerEquity,
+        }
+        const dividendData_3 = {
+            recordDate: dividendsRecordDateInSeconds_3.toString(),
+            executionDate: dividendsExecutionDateInSeconds.toString(),
+            amount: dividendsAmountPerEquity,
+        }
+        await equityFacet.setDividends(dividendData_2)
+        await equityFacet.setDividends(dividendData_3)
+        await equityFacet.setDividends(dividendData_1)
 
-        const balanceAdjustmentData_1 = {
-            executionDate: balanceAdjustmentExecutionDateInSeconds_1.toString(),
-            factor: balanceAdjustmentsFactor,
-            decimals: balanceAdjustmentsDecimals,
-        }
-        const balanceAdjustmentData_2 = {
-            executionDate: balanceAdjustmentExecutionDateInSeconds_2.toString(),
-            factor: balanceAdjustmentsFactor,
-            decimals: balanceAdjustmentsDecimals,
-        }
-        const balanceAdjustmentData_3 = {
-            executionDate: balanceAdjustmentExecutionDateInSeconds_3.toString(),
-            factor: balanceAdjustmentsFactor,
-            decimals: balanceAdjustmentsDecimals,
-        }
-        await equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData_2)
-        await equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData_3)
-        await equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData_1)
-
-        const balanceAdjustment_2_Id =
+        const dividend_2_Id =
             '0x0000000000000000000000000000000000000000000000000000000000000001'
-        const balanceAdjustment_3_Id =
+        const dividend_3_Id =
             '0x0000000000000000000000000000000000000000000000000000000000000002'
-        const balanceAdjustment_1_Id =
+        const dividend_1_Id =
             '0x0000000000000000000000000000000000000000000000000000000000000003'
 
-        // check schedled BalanceAdjustments
-        scheduledBalanceAdjustmentsFacet =
-            scheduledBalanceAdjustmentsFacet.connect(signer_A)
+        // check schedled snapshots
+        scheduledSnapshotsFacet = scheduledSnapshotsFacet.connect(signer_A)
 
-        let scheduledBalanceAdjustmentCount =
-            await scheduledBalanceAdjustmentsFacet.scheduledBalanceAdjustmentCount()
-        let scheduledBalanceAdjustments =
-            await scheduledBalanceAdjustmentsFacet.getScheduledBalanceAdjustments(
-                0,
-                100
-            )
+        let scheduledSnapshotCount =
+            await scheduledSnapshotsFacet.scheduledSnapshotCount()
+        let scheduledSnapshots =
+            await scheduledSnapshotsFacet.getScheduledSnapshots(0, 100)
 
-        expect(scheduledBalanceAdjustmentCount).to.equal(3)
-        expect(scheduledBalanceAdjustments.length).to.equal(
-            scheduledBalanceAdjustmentCount
+        expect(scheduledSnapshotCount).to.equal(3)
+        expect(scheduledSnapshots.length).to.equal(scheduledSnapshotCount)
+        expect(scheduledSnapshots[0].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_3
         )
-        expect(
-            scheduledBalanceAdjustments[0].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_3)
-        expect(scheduledBalanceAdjustments[0].data).to.equal(
-            balanceAdjustment_3_Id
+        expect(scheduledSnapshots[0].data).to.equal(dividend_3_Id)
+        expect(scheduledSnapshots[1].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_2
         )
-        expect(
-            scheduledBalanceAdjustments[1].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_2)
-        expect(scheduledBalanceAdjustments[1].data).to.equal(
-            balanceAdjustment_2_Id
+        expect(scheduledSnapshots[1].data).to.equal(dividend_2_Id)
+        expect(scheduledSnapshots[2].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_1
         )
-        expect(
-            scheduledBalanceAdjustments[2].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_1)
-        expect(scheduledBalanceAdjustments[2].data).to.equal(
-            balanceAdjustment_1_Id
-        )
+        expect(scheduledSnapshots[2].data).to.equal(dividend_1_Id)
 
-        // AFTER FIRST SCHEDULED BalanceAdjustmentS ------------------------------------------------------------------
+        // AFTER FIRST SCHEDULED SNAPSHOTS ------------------------------------------------------------------
         scheduledTasksFacet = scheduledTasksFacet.connect(signer_A)
+
         await timeTravelFacet.changeSystemTimestamp(
-            balanceAdjustmentExecutionDateInSeconds_1 + 1
+            dividendsRecordDateInSeconds_1 + 1
         )
-        await scheduledTasksFacet.triggerPendingScheduledTasks()
+        await expect(scheduledTasksFacet.triggerPendingScheduledTasks())
+            .to.emit(scheduledSnapshotsFacet, 'SnapshotTriggered')
+            .withArgs(account_A, 1)
 
-        scheduledBalanceAdjustmentCount =
-            await scheduledBalanceAdjustmentsFacet.scheduledBalanceAdjustmentCount()
-        scheduledBalanceAdjustments =
-            await scheduledBalanceAdjustmentsFacet.getScheduledBalanceAdjustments(
-                0,
-                100
-            )
+        scheduledSnapshotCount =
+            await scheduledSnapshotsFacet.scheduledSnapshotCount()
+        scheduledSnapshots =
+            await scheduledSnapshotsFacet.getScheduledSnapshots(0, 100)
 
-        expect(scheduledBalanceAdjustmentCount).to.equal(2)
-        expect(scheduledBalanceAdjustments.length).to.equal(
-            scheduledBalanceAdjustmentCount
+        expect(scheduledSnapshotCount).to.equal(2)
+        expect(scheduledSnapshots.length).to.equal(scheduledSnapshotCount)
+        expect(scheduledSnapshots[0].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_3
         )
-        expect(
-            scheduledBalanceAdjustments[0].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_3)
-        expect(scheduledBalanceAdjustments[0].data).to.equal(
-            balanceAdjustment_3_Id
+        expect(scheduledSnapshots[0].data).to.equal(dividend_3_Id)
+        expect(scheduledSnapshots[1].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_2
         )
-        expect(
-            scheduledBalanceAdjustments[1].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_2)
-        expect(scheduledBalanceAdjustments[1].data).to.equal(
-            balanceAdjustment_2_Id
-        )
+        expect(scheduledSnapshots[1].data).to.equal(dividend_2_Id)
 
-        // AFTER SECOND SCHEDULED BalanceAdjustmentS ------------------------------------------------------------------
+        // AFTER SECOND SCHEDULED SNAPSHOTS ------------------------------------------------------------------
         await timeTravelFacet.changeSystemTimestamp(
-            balanceAdjustmentExecutionDateInSeconds_2 + 1
+            dividendsRecordDateInSeconds_2 + 1
         )
-        await scheduledTasksFacet.triggerScheduledTasks(100)
+        await expect(scheduledTasksFacet.triggerScheduledTasks(100))
+            .to.emit(scheduledSnapshotsFacet, 'SnapshotTriggered')
+            .withArgs(account_A, 2)
 
-        scheduledBalanceAdjustmentCount =
-            await scheduledBalanceAdjustmentsFacet.scheduledBalanceAdjustmentCount()
-        scheduledBalanceAdjustments =
-            await scheduledBalanceAdjustmentsFacet.getScheduledBalanceAdjustments(
-                0,
-                100
-            )
+        scheduledSnapshotCount =
+            await scheduledSnapshotsFacet.scheduledSnapshotCount()
+        scheduledSnapshots =
+            await scheduledSnapshotsFacet.getScheduledSnapshots(0, 100)
 
-        expect(scheduledBalanceAdjustmentCount).to.equal(1)
-        expect(scheduledBalanceAdjustments.length).to.equal(
-            scheduledBalanceAdjustmentCount
+        expect(scheduledSnapshotCount).to.equal(1)
+        expect(scheduledSnapshots.length).to.equal(scheduledSnapshotCount)
+        expect(scheduledSnapshots[0].scheduledTimestamp.toNumber()).to.equal(
+            dividendsRecordDateInSeconds_3
         )
-        expect(
-            scheduledBalanceAdjustments[0].scheduledTimestamp.toNumber()
-        ).to.equal(balanceAdjustmentExecutionDateInSeconds_3)
-        expect(scheduledBalanceAdjustments[0].data).to.equal(
-            balanceAdjustment_3_Id
-        )
+        expect(scheduledSnapshots[0].data).to.equal(dividend_3_Id)
 
-        // AFTER THIRD SCHEDULED BalanceAdjustmentS ------------------------------------------------------------------
+        // AFTER SECOND SCHEDULED SNAPSHOTS ------------------------------------------------------------------
         await timeTravelFacet.changeSystemTimestamp(
-            balanceAdjustmentExecutionDateInSeconds_3 + 1
+            dividendsRecordDateInSeconds_3 + 1
         )
-        await scheduledTasksFacet.triggerScheduledTasks(0)
+        await expect(scheduledTasksFacet.triggerScheduledTasks(0))
+            .to.emit(scheduledSnapshotsFacet, 'SnapshotTriggered')
+            .withArgs(account_A, 3)
 
-        scheduledBalanceAdjustmentCount =
-            await scheduledBalanceAdjustmentsFacet.scheduledBalanceAdjustmentCount()
-        scheduledBalanceAdjustments =
-            await scheduledBalanceAdjustmentsFacet.getScheduledBalanceAdjustments(
-                0,
-                100
-            )
+        scheduledSnapshotCount =
+            await scheduledSnapshotsFacet.scheduledSnapshotCount()
+        scheduledSnapshots =
+            await scheduledSnapshotsFacet.getScheduledSnapshots(0, 100)
 
-        expect(scheduledBalanceAdjustmentCount).to.equal(0)
-        expect(scheduledBalanceAdjustments.length).to.equal(
-            scheduledBalanceAdjustmentCount
-        )
+        expect(scheduledSnapshotCount).to.equal(0)
+        expect(scheduledSnapshots.length).to.equal(scheduledSnapshotCount)
     })
 })
